+++ compute/proto_compute.js
@@ -380,8 +380,11 @@ steal('can/util', 'can/util/bind', 'can/compute/read.js','can/compute/get_value_
 	// ## updateOnChange
 	// A helper to trigger an event when a value changes
 	var updateOnChange = function(compute, newValue, oldValue, batchNum){
+
+		var valueChanged = newValue !== oldValue && !(newValue !== newValue && oldValue !== oldValue);
+		
 		// Only trigger event when value has changed
-		if (newValue !== oldValue) {
+		if (valueChanged) {
 			can.batch.trigger(compute, {type: "change", batchNum: batchNum}, [
 				newValue,
 				oldValue