+++ view/parser/parser.js
@@ -26,8 +26,10 @@ steal(function(){
 	var alphaNumericHU = "-:A-Za-z0-9_",
 		attributeNames = "[^=>\\s\\/]+",
 		spaceEQspace = "\\s*=\\s*",
+		singleCurly = "\\{[^\\}\\{]\\}",
+		doubleCurly = "\\{\\{[^\\}]\\}\\}\\}?",
 		attributeEqAndValue = "(?:"+spaceEQspace+"(?:"+
-		  "(?:\"[^\"]*\")|(?:'[^']*')|[^>\\s]+))?",
+		  "(?:"+doubleCurly+")|(?:"+singleCurly+")|(?:\"[^\"]*\")|(?:'[^']*')|[^>\\s]+))?",
 		matchStash = "\\{\\{[^\\}]*\\}\\}\\}?",
 		stash = "\\{\\{([^\\}]*)\\}\\}\\}?",
 		startTag = new RegExp("^<(["+alphaNumericHU+"]+)"+
@@ -287,6 +289,7 @@ steal(function(){
 		state.valueStart = undefined;
 		state.inValue = false;
 		state.inName = false;
+		state.lookingForEq = false;
 		state.inQuote = false;
 		state.lookingForName = true;
 	};
@@ -315,8 +318,7 @@ steal(function(){
 			var next = rest.charAt(i+1);
 			var nextNext = rest.charAt(i+2);
 			i++;
-			
-			
+			//debugger;
 			if(cur === "{" && next === "{") {
 				if(state.inValue && curIndex > state.valueStart) {
 					handler.attrValue(rest.substring(state.valueStart, curIndex));
@@ -324,6 +326,14 @@ steal(function(){
 				// `{{#foo}}DISABLED{{/foo}}`
 				else if(state.inName && state.nameStart < curIndex) {
 					callAttrStart(state, curIndex, handler, rest);
+					callAttrEnd(state, curIndex, handler, rest);
+				}
+				// foo={{bar}}
+				else if(state.lookingForValue){
+					state.inValue = true;
+				}
+				// a {{bar}}
+				else if(state.lookingForEq && state.attrStart) {
 					callAttrEnd(state, curIndex, handler, rest);
 				}
 				state.inDoubleCurly = true;
@@ -398,6 +408,8 @@ steal(function(){
 			callAttrStart(state, curIndex+1, handler, rest);
 			callAttrEnd(state, curIndex+1, handler, rest);
 		} else if(state.lookingForEq) {
+			callAttrEnd(state, curIndex+1, handler, rest);
+		} else if(state.inValue) {
 			callAttrEnd(state, curIndex+1, handler, rest);
 		}