+++ view/stache/stache_test.js
@@ -4603,6 +4603,27 @@ steal("can-simple-dom", "can/util/vdom/build_fragment","can/view/stache", "can/v
 			var frag = can.stache("<div {{#if preview}}checked{{/if}}></div>")(map);
 			equal(frag.firstChild.getAttribute("checked"), "", "got attribute");
 		});
+		
+		test("sections with attribute spacing (#2097)", function(){
+			var template = can.stache('<div {{#foo}} disabled {{/foo}}>');
+			var frag = template({foo: true});
+			equal(frag.firstChild.getAttribute("disabled"),"","disabled set");
+		});
+		
+
+		test("keep @index working with multi-dimensional arrays (#2127)", function() {
+			var data = new can.Map({
+				array2 : [['asd'], ['sdf']]
+			});
+			
+			var template = can.stache('<div>{{#each array2}}<span>{{@index}}</span>{{/each}}</div>');
+			
+			var frag = template(data);
+			
+			var spans = frag.firstChild.getElementsByTagName("span");
+			equal( spans[0].firstChild.nodeValue, "0");
+		});
+
 		// PUT NEW TESTS RIGHT BEFORE THIS!
 	}